Fixing broken validation tests.
There was some interesting fallout from an earlier change (outside of this PR) where a field that was incorrectly encoded as a string was now being correctly encoded as a number.

Unfortunately our message validation code was _not_ running on the message if this incorrect code path was not triggered, leading to us falling back to the service instead (unintentional).

I've created a unit test suite for the message, moved the valiation code into the primary function called (toRheaMessage). So this error shouldn't happen again.