[service-bus] Fixing session peeking test that was failing randomly. (#14872)
This commit adds in a receiveMessages call to ensure that messages have arrived by the time we actually do our peekMessages() call. Without this, if there is a delay between the message being processed in the queue and it being available for receiving we'll end up in a situation where peekMessages() will return zero messages.

peekMessages goes against an expected sequence number and doesn't wait if there aren't any messages at that moment to receive. receiveMessages() always asks for the current messages and will wait for the expected number to arrive. So by adding peek _after_ receiveMessages() we should be able to get the proper sequence of operations - messages arrive, are available and now we can peek properly.

Fixes #14757