[Identity] Reverting ClientSecretCredential and UsernamePasswordCredential to v1 only for browsers, and fixing Karma browser bundles (#14910)
This PR makes it so `ClientSecretCredential` behaves like how it did on v1 only for browsers. It also ensures browser bundles can be used when built with Karma.

---

## ClientSecretCredential (and also UsernamePasswordCredential)

Since the inception of Identity, we've been relying on `ClientSecretCredential` to be available in browsers to make it easier for us to automate bowser tests.

Whether `ClientSecretCredential` should continue being available on browsers in the future, is something yet to be decided. We started a conversation on the matter here: #14879 

Our current plan is to switch master back to v1, as you can see here: #14909

Since the v2 changes to `ClientSecretCredential` won't be reverted, that means that as soon as we move master back to v1, all of our current clients will immediately start working with the latest `ClientSecretCredential`.

`ClientSecretCredential` was rewritten to use MSAL, specifically `@azure/msal-node`, which is not intended to work on browsers.

This PR makes it so `ClientSecretCredential` is kept as it was on v1 only for browsers. This will:

- Allow us to keep our current browser recordings.
- Allow us to continue using this credential for browser tests.
- On v1, dropping this credential on browsers would be a breaking change, so this technically reduces the differences with v1.

**Important:**
Keep in mind that while we've always supported this credential in browsers, it can only work if browser security is disabled. This isn't recommended outside of our tests, and it's not something we should be advertising.

(Later we also did the same to `UsernamePasswordCredential`).

---

## Browser bundles

While my initial change was scoped to one credential, I asked Harsha to test this PR and he found that the browser bundles would not work with Identity v2. We decided to fix it right away since this PR wouldn't be very useful with broken browser bundles.