[Identity] Updating msal-node, and re-enabling the DeviceCodeCredential cancel() test (#14881)
MSAL just released `@azure/msal-node@^1.0.2`, which fixes the cancelation of pending requests. This PR updates this dependency and re-enables the test.

Device code is a special case because it has to do polling. Most credentials send network requests in sequence, meaning that when there's a delay, it comes from the service, and thus cancelation through the abort-controller is feasible. We have abort-controller tests for ClientSecretCredential, ClientCertificateCredential and UsernamePasswordCredential, for example.

Note that, since the polling interval that MSAL uses is defined by the values it receives from the Identity endpoint, on playback we replace the polling interval from 5s to 0.1s, making this test fail. I tried tweaking the delay, but it's very fragile on playback. This test will pass during recording, and on live tests.

Fixes #14376