[KeyVault] - Regenerate from 7.2 stable swagger (#14964)
This PR includes regenerated code from the stable 7.2 swagger. Up until now it was in preview, but now that it's 
stable we're regenerating (and expected no meaningful changes).

But there were meaningful changes! Recently our code generator started emitting `null` based on the x-nullable spec
which unfortunately expanded CertificateOperation.error and ErrorModel.innerError to be `ErrorModel | undefined | **null**`
Since this data is going from server to client we can only narrow it, we can't expand it without a breaking change. This is 
made worse because we re-exported ErrorModel as-is in 4.1. 

I looked into it and realized that this is done for historical reasons and not semantically meaningful 
see: https://github.com/Azure/azure-rest-api-specs/pull/10262#discussion_r461133511.

So I am deprecating `ErrorModel` since we should use the `CertificateOperationError` type anyway, ensuring that the 
existing ErrorModel does not change, and applying the transformation from null to undefined in the convenience layer to 
keep the API the same.