Add credential decision matrix in AzureIdentityExamples.md (#14920)
This PR adds the tables for different auth mechanisms we have in `@azure/identity` to the AzureIdentityExamples.md file. This is inspired from the content in the readme file of  `@azure/identity`, but augmented to have a bit more description and include notes on set up steps on has to go through before using the credential.

I have taken the liberty to restructure the headers in this file as the table now gives a better way to navigate than rely on a list that has more than 10 entries making it difficult to skim.

If we are happy with this content, we can pull the tables into a separate md file and use the `include` notation to include it in AzureIdentityExamples.md as well as the readme file.