Fixing quite a few tests. Some of them were just being a bit too chummy with internal functions. For the most part though the tests got simpler since there were fewer functions that needed mocking (subscribe now does initialization and subscription), etc..
There were also some interesting breakages when I changed assertThrow to chain on the promises returned (which made my test callstacks a bit nicer) in that one of the methods that's being tested (settleMessages) was not marked as async, causing it to throw exceptions that were not "catchable" by chaining to the returned promise.

I've also enabled code coverage, even for unit tests. They go under the coverage/unit-tests folder, just to keep it all clean.