[core-rest-pipeline] WWW-Authenticate challenge perf-stress test (#14659)
This PR is a draft of how the WWW-Authenticate perf-stress would look.

**What this first performance test does:**  

This initial test only sets up a pipeline that does a service request against a fake server that answers once with a WWW-Authenticate challenge, and the next time with a 200. So, the first request will fail, we'll get a challenge, the challenge will be parsed and another request will be sent afterwards, which will succeed.

To run the test, we call to: `npm run perf-test:node -- BearerTokenChallengeAuthenticationPolicyTest --warmup 1 --iterations 1 --parallel 5`.

Here's how it looks when it runs:

```
=== Calling globalSetup() once for (all) the instance(s) of BearerTokenChallengeAuthenticationPolicyTest ===

=== warmup mode, iteration 1. Logs every 1s ===
Current         Total           Average
=== warmup mode, results of iteration 1 ===
Completed 55,001 operations in a weighted-average of 1.00s (55,015.93 ops/s, 0.000 s/op)

=== test mode, iteration 1. Logs every 1s ===
Current         Total           Average
=== test mode, results of iteration 1 ===
Completed 537,426 operations in a weighted-average of 10.00s (53,741.61 ops/s, 0.000 s/op)
```

**Background:**  

Java's perf-stress for challenges uses the ARM kind of challenges, but on TypeScript our more immediate customer is not ARM, but Container Registry and then Key Vault.

This PR uses a copy of the challenge test we have in the main challenge PR: #13888 (this PR hasn't been merged, the plan is to have it merged by next week). Keep in mind that this PR won't pass CI until #13888 is merged.

The idea with this PR is that, besides perf-stress testing a basic scenario, we can use it as the base for further perf-stress tests for core-rest-pipeline.

Fixes #14633