Fixing a test that was a bit brittle looking into something that relies more on public APIs.
It did point out a simple use-case where we shouldn't bug the user with an error - if they suspend the receiver (and thus addCredit throws AbortErrors) we can ignore those since the user has initiated the action that caused it and there'd be nothing useful about reporting that specific error.

Also, I made the session "failed to add credit" message a bit more like the one we throw in non-session receivers.