[Dev tool] Support samples v2 directory structure in check node versions command (#15328)
Samples v2 uses a directory structure like this `samples/v<some package version>` instead of the good old `samples`. Furthermore, it is not guaranteed that the package version is the most recent major one because a package author can decide to change it to another number. This PR adds logic to handle both samples v1 and v2 and make sure to use the appropriate corresponding path. If there is more than one samples v2 directory, the command will run the samples in the directory with the largest version number.