[identity] Unpin msal-common and msal-browser dependencies (#15338)
Dependency analysis was showing that we are pulling in multiple copies of `msal-common` in the install footprint of identity. We were minor-pinned to msal-common 4.0.x (latest is 4.3.0) and hard pinned to a specific version of msal-browser. However, msal-browser itself carries a dependency on msal-common, and so our install footprint is currently pulling in _three_ different versions of msal-common. Letting these dependencies roll forward within the same major version should help.

![image](https://user-images.githubusercontent.com/77019085/118858126-a4d11f00-b88d-11eb-8ee6-db893623a627.png)

CC @HarshaNalluru @sadasant -- is this something that will give us recorder pain?