[Key Vault Keys] Fixed bad references. (#15360)
A customer recently found an issue with our output values: https://github.com/Azure/azure-sdk-for-js/issues/15353

Once we reach an exit strategy, we'll need to make a hotfix out of this PR.

A while ago I made a PR showcasing how we could use the spread operator less ([PR link](https://github.com/Azure/azure-sdk-for-js/pull/12697)). That PR got reviewed and approved as usual, but it had some deep issues that didn't become obvious to the reviewer, nor raised any issues during tests. I find that reasonable, but I regret it.

What happened in detail was that, because of:

- The use of "any".
- Not having structural tests.

Because this being an instance of both cases, we couldn't:

- Trust our types.
- Trust our tests.

So, besides solving the underlying issue, we should take this opportunity to reflect on the matter.

Feedback appreciated.