[EventGrid] Updates to System Event Mappings
This change updates to the latest version of the EventGrid data plane
specification, and brings in some new event types (one for Storage and
a handful for FarmBeats).

In addition there is a small bugfix to one of the event names
(discovered as we move towards tooling the process that generates
these mappings) to use the correct name.