[Identity] Different approach for testing the Arc feature, instead of mock-fs (#15571)
Recently we found an issue on our CI build for Node 16.3.0, which boiled down to one unit test specific to Identity: https://github.com/Azure/azure-sdk-for-js/pull/15547

After some investigation, it is clear that Node 16.3.0 has some incompatibility with mock-fs: https://github.com/tschaub/mock-fs/issues/332

In the mean time, to add this test again, we can simply create a file, point to it and delete it at the end. This should work in all versions of Node, and it means one less dev-dependency for us.