[KeyVault] - Return void from delete methods (#15642)
## What

-KeyVaultAccessControlClient#deleteRoleAssignment changes to return `Promise<void>` instead of `Promise<RestResponse>`
-KeyVaultAccessControlClient#deleteRoleDefinition changes to return `Promise<void>` instead of `Promise<RestResponse>`

## Why

Originally we wanted to return the raw response so that customers can access the headers / status code / etc. However, with the 
upcoming effort to core-v2 returning a raw response today would require us to make a breaking change tomorrow as corev2 
does not return the raw response, using a callback instead.

By returning `Promise<void>` today we will be ready to change to corev2 tomorrow without introducing a breaking change to 
these methods.