[EventGrid] Move to Core V2 Generator
As part of the development of the new pipeline, event grid was hand
ported to use the new pipeline. Now that the code generator targets
the new pipeline, we can start using it to generate the code.

Since EventGrid includes the `/api/events` path segement in the
Endpoint, We need to do a small amount of post processing of the
generated code, to ensure `/api/events` is not appended to the
endpoint (we do this by setting an empty path in the operation spec,
which is as things were before moving over to the generator).

Fixes #15823