[KeyVault] - Fix incorrect paths to keyvault-common (#15751)
## What

-Ensure all imports to keyvault-common go through the src directory

## Why

We noticed that rollup failed to find imports from keyvault common when we updated our test dependencies to use the published KV Keys version. 

This is not an issue when we were using the version of KV Keys on master because of the way rush would use symlinks; however,
this is an issue in a certain edge case and therefore should be fixed,

Longer term we'd like to run at least one of our packages against the published version to find issues like this sooner.