[Core] - Move core-tracing to @opentelemetry/api 0.20.0 (#15672)
## What

- Bump @opentelemetry/api to 0.20.0 in @azure/core-tracing
- Move all packages that are on core-http to the next core-tracing version 
- Remove version collision check from the tracer cache

## Why

This is part of our effort to move everyone to OTel 0.20.0 - but we have to stage it due to a transitive dependency and a 
breaking change in OTel. This PR updates core-tracing to use the latest OTel, fixes any breaking changes, and moves 
packages that we can move to the latest version of core-tracing.

Once core-rest-pipeline 1.1.0 is GA'd we'll be able to move the rest of the packages over to the latest core-tracing as well.

Removing the version collision came out of an offline discussion after these changes were reviewed - since it's a common 
source of pain and hasn't added much benefit (tracer conflicts incompatibility was never a problem) we decided to remove 
that logic and always grab a unique symbol per OT compatibility. 

## Callouts

The packages that are already on core-v2 will _not_ be upgraded at this time - we'll coordinate that with the GA of core-rest-pipeline 1.1.0