[Identity] Comment improvements. Identity generally doesn't return null anymore (#15783)
I’ve got some customers writing to me directly with some confusion regarding getToken returning null. We’ve got some remnants of this type, but at this point, Identity doesn’t return null (except for ClientSecretCredential in the browser, which is generally unusable outside of our tests). This PR is to update these docs and to give a bit of clarity on what errors we’re throwing.