[synapse] - Ignore known rollup warnings (#15948)
## What

- Adds warning filter to ignore known rollup warnings that are safe to ignore
- Ignore `this has been rewritten to undefined` in synapse
- Ignore circular dependencies in synapse

## Why

These are both known warnings, are safe to ignore, and add noise to the build output unnecessarily. 

The first is totally safe to ignore:

```
var __spreadArray = (this && this.__spreadArray) || function (to, from) {
```

Is emitted from TypeScript in the OTel ESM and is safe to use.
The second is a known issue in OTel 1.0.0 that is not harmful at runtime.

Finally, we have precedent for doing the exact same thing in other packages like service-bus.