Removing three tests which had more code than needed to test that the maxConcurrentCalls was passed properly.
Replaced with a smaller test that can (with the right parameters, manually specified) still cause the session overlap bug. I don't have a solution for that one yet but this will allow us to investigate it when we're ready.

Part of the investigation for https://github.com/Azure/azure-sdk-for-js/issues/14441