[Identity] [Hotfix] 1.5.0: Migrated from core-http to core-rest-pipeline (#16432)
This PR aims to:

- Migrate Identity 1.4.0 from `@azure/core-http` to `@azure/core-rest-pipeline` for the handling of HTTP requests.
- Enabled Service Fabric support on the `ManagedIdentityCredential`.

Service Fabric was disabled because we we relying on core-http, which tried to verify the HTTPS certificate. I believe that this won’t be the case with the new core if `allowInsecureConnection` is set to true, is that correct? @xirzec - my mocked tests passed, but I haven’t tested this on the actual Service Fabric.

This is relevant to Karishma since she has been working on the e2e tests for Identity on Service Fabric (… turns out it wasn’t enabled! I had forgotten). @KarishmaGhiya please help me test these changes on Service Fabric 🙇‍♂️ let’s sync up today and try this together.

If it isn’t as straightforward as I thought it would be, I can drop ServiceFabric support again 😄 that takes two lines of code (well, 1 line of code and the changelog line).

I think 1.5.0 makes sense for this change, given that we’re swapping the network adapter.