[event-hubs] prevent invalid spans from instrumenting EventData (#16439)
The live tests for event hubs currently have 3 test failures due to some changes introduced in core-tracing preview 13. The changes to core-tracing cause EventData to be instrumented (via Diagnostic-Id) with invalid span context data.

This change adds a check to ensure the span context is valid before instrumenting the EventData to be sent to the service.

Note: I didn't add a changelog entry because from a user's point of view, nothing has changed, since this issue did not exist in the previously released version of the event-hubs package.