Delete mgmt pipeline for Storage hotfix branch (#16543)
The PR #16529 for a hotfix for Storage packages is currently blocked due to failure in the `js - mgmt - pr` pipeline. This pipeline should not have even been triggered for this PR. This pipeline is for Track 1 auto generated packages (both data plane & mgmt plane) and this PR does not touch any of those packages.

Since we need to prep the hotfix for the storage team to release this Monday, this PR deletes this pipeline in the hotfix branch as a stop gap. 

Logged #16542 for follow up as to why this pipeline was triggered