[KeyVault] - Fix double-purge of the same key in test (#16547)
After a month of not being able to run live tests due to a service performance issue, we have finally been able to run live 
Key Vault tests. The live test run showed a single failure, which is caused by a test purging a key which later is purged by the
afterEach hook as well. The second purge would fail, since the key doesn't exist.