[Tables] Fix batch operations against emulators
In #15938 we exposed the `allowInsecureConnection` option to allow
connections to the service over HTTP instead of HTTPS. However, even
if a table client was created allowing insecure connections, we would
still require a secure connection when submitting a batch transaction,
since the implementation of batch operations interact with the
pipeline directly. We now record if the client was created allowing
insecure connections and thread that information along when creating
requests for batch operations.

With this change, you can now submit batch transactions to storage
emulators, which commonly run over HTTP instead of HTTPS.

Fixes #15854