[KeyVault] - Update core libraries for KeyVault Admin (#16581)
Now that @opentelemetry/api has gone GA, we have more users attempting to try it. The latest stable releases for keyvault packages still use an older version of @azure/core-tracing that depends on 1.0.0-rc version of @opentelemetry/api. This results in problems like the one seen in #16447

Moreover, our recent move to use 2.0.0 of @azure/core-http results in different versions of core-http being used in cases like #16389 and failing builds even when the user does not use tracing.

Key Vault Admin is currently in the beta phase of the next release. 

To resolve such issues in the near term, we would like to have a stable release out for the keyvault packages that uses the latest @azure/core-tracing package. This PR updates keyvault admin to a version of core-tracing that depends on OTel ^1.0.0 and will be merged into hotfix/keyvault-admin-opentelemetry-ga