[EventGrid] Fix Test Compliation Errors
In #16700 we updated our tests to generate the key we used for testing
by doing a base64 conversion of a fixed string so we didn't have something that looked like a key embeded in source which was upsetting some static analysis tools.

This casues a build break in the min/max testing where `btoa` can't be found when building the test. Since we have the correct runtime behavior in the test to only call `btoa` when we are not in a node-context we simply reference the dom library to appease the compiler.