[KeyVault] - Add releasePolicy to UpdateKeyProperties (#16722)
## What

- Add `releasePolicy` to `UpdateKeyPropertiesOptions`
- Add an overload to `updateKeyProperties` that does not take a key version

## Why

While you are unable to update the exportable attribute of a key once it's been created, you should be able to update the 
release policy of an exportable key; however, we have a bug in the GA version of KV Keys where `updateKeyProperties` 
requires a version. While that has not come up from customers yet, and a workaround is to pass something falsey for key 
version, it feels incorrect that we are requiring an argument that is functionally optional.

This wasn't necessary until now, but version _must_ be omitted when updating a key's release policy, so this is now a 
requirement.

Normally, the version is part of the `<operation>Options` bag, but in this case an overload felt more appropriate although 
inconsistent because moving the version param to the options bag would be a breaking change for consumers. 

Resolves #16721