[Tables] Infer EDM types for properties with no Type metadata (#16736)
**Context:**
When querying the Tables service we have an option to disable automatic type conversion. This means that we would give users an object similar to what the service sends us. For example, an object with a Date property will be received like this, with an `@odata.type` metadata property

```json
{
   "PartitionKey": "1",
   "RowKey": "1",
   "dateValue": "2020-09-17T00:00:00.111Z",
   "dateValie@odata.type": "Edm.DateTime"
}
```

The SDK's automatic type conversion would use the metadata to give users a primitive Date type so the user would get the following object

```javascript
{
   partitionKey: "1",
   rowKey: "1",
   dateValue: new Date("2020-09-17T00:00:00.111Z")
}
```

If `disableTypeConversion` is true in the list or get operations the sdk returns the Edm object instead of a primitive type

```javascript
{
   partitionKey: "1",
   rowKey: "1",
   dateValue: {value: "2020-09-17T00:00:00.111Z", type: "DateTime"}
}
```

**Problem**
There are 3 types when the service never sends `@odata.type` metadata and just the value, these are int32, double, and string. Before this PR we would return the value instead of the EDM object since no @odata.type property was found. This causes an inconsistent experience.

**Fix**
When `disableTypeConversion` is true, return an EDM object instead of the value by inferring the EDM type, in the case of the string is straightforward. In the case of int32 and decimal, JavaScript treats both as `number` so we need to inspect the value looking for decimal places to decide whether to use `Edm.Int32` or `Edm.Decimal`.


**Note:**
There is a limitation with this approach, and it is that JavaScript drops decimals if they are `0` for example, `1.00` would be represented as `1`. With these changes, such cases would always have an Edm type of `Int32`