[KeyVault] - Increase timeout for KeyVault certificates and secrets (#16782)
## What

- Bump the timeout of integration tests for KV Secrets and Certificates
- Bump the timeout of min/max testing to work around #16731 
- Delete extra cleanup methods 

## Why

Due to an ongoing performance regression on the service side our tests have been timing out. The performance has gotten better but after talking to the service team it sounds like the recommendation is to avoid waiting on deletes / purges if you
are latency sensitive. 

However, not all our tests are just cleaning up, there are valid tests that exercise cleaning / purging certificates so bumping
the timeout a bit feels appropriate. This works fine for tests except that min/max testing uses a hardcoded number for test
timeouts so need to update that as well (or add timeouts everywhere in our code). Since certificates and secrets _already_
has the longest timeout period bumping it in min/max seems appropriate until #16731 is resolved.