[storage] - Remove unnecessarily extending CommonOptions (#16801)
## What

- change BlobBatchSubmitBatchOptionalParams to no longer extend CommonOptions
- remove the declaration of `abortSignal` in BlobBatchSubmitBatchOptionalParams

## Why

BlobBatchSubmitBatchOptionalParams already extends coreHttp.OperationOptions indirectly through ServiceSubmitBatchOptionalParams and does not need to pull tracingOptions from CommonOptions anymore. It also doesn't need to re-declare abortSignal as it gets that from OperationOptions as well.

So nothing changes, but we also know that enums which come in from tracingOptions are compatible but not identical, so when extending both TS will throw an error if core-tracing versions are different. See #15285

This change resolves the above in the only known case of this in our repo, cleans up the unnecessary duplication, and does not change the public API.