[Identity] Test improvements (#16821)
`path.posix.join` changes `https://` to `https:/`. That happens to be automatically fixed by our core, so the tests didn’t break before. This PR ensures we always send a reasonable URL to the core.

We also test with a non-default environment, to make sure the tests don’t accidentally use the default values.

Thanks to @chlowell for the feedback!