[KeyVault] - Delete and purge KV and MHSM after live tests (#16819)
## What

- Add a script to find, delete, and purge any key vaults or managed HSMs within the test resource group

## Why

Quotas and costs - both Key Vault and Managed HSM have soft delete enabled by default (which cannot be disabled). This
means that deleting the resource group is not sufficient, and purging the resource is necessary in order to truly delete it.

We already set the softDeleteRetentionInDays parameter to 7 days (the minimum allowed) but with all the live test runs we 
are still collecting a large number of resources in a soft-deleted state.

For KV - We will soon have a hard limit of 1000 soft deleted keyvaults per subscription and we are at a much higher number 
than that today.

For MHSM - There is a 5 MHSM per region per subscription limit which we will easily exceed if we don't purge them.

## Callouts

There are no PowerShell Cmdlets that let you purge a soft deleted MHSM yet, so until they're available I am using the 
Azure CLI like we did during deployment.