[KeyVault] - Add support for Key Rotation (#16593)
## What

- Add `getKeyRotationPolicy`, `rotateKey`, and `updateKeyRotationPolicy` to KeyClient
- Model Key Rotation policy similarly to Certificate Policy
- Add `rotate` permission to the arm template when deploying keyvault

## Why

As part of 7.3 the KV crew is adding support for Key Rotation which we will need to model in our SDK. It is currently in preview
but is planned for 7.3. As such, we have a few methods to model here:

- Updating a key's rotation policy
- Fetching a key's rotation policy
- Rotating a key on-demand which will create a new version of the key

In addition, since it requires the `/keys/rotate` permission, we need to add that to the list of permissions in our test resources. 

## Callout

As of 8/18/21 the model for rotation policy contains ISO8601 durations which we are modeling as string similar to our other
libraries. That might change at the swagger level or it might change at the SDK level in the near future (decision TBD)