[ContainerRegistry] Update cloud configuration API
We are moving from an `authenticationScope` string to an `audience`
extensible enum on `ContainerRegistryClientOptions`. We introduce the
`KnownContainerRegistryAudience` enum to hold the possible values, or
allow the customer to provide a different value if needed.

In addition, we require the `audience` value to be set in client
options to instantiate any instance of `ContainerRegistryClient`. This
allows us to postpone setting a default value for audience until after
the library's first GA.

See relevant .NET PR at
https://github.com/Azure/azure-sdk-for-net/pull/23141.

This resolves https://github.com/azure/azure-sdk-for-js/issues/16127.