[Identity] Throwing if required parameters are not provided (#16846)
Since I moved most of the processing to a common place (`nodeCommon.ts`), some of the requirements per credential were obscured. `ClientSecretCredential` and `ClientCertificateCredential` in particular where not throwing early if the parameters weren’t provided (which of course is possible if users bypass the parameter types, or if they use JavaScript). I’ve also added tests to confirm this behavior in the future.