Update CODEOWNERS for every mgmt package in feature/v4 branch (#17061)
As discussed in https://github.com/Azure/azure-sdk-tools/issues/1099, usage of glob patterns does not work for our bot. So adding codeowner for each of the mgmt plane package. I copied the list from https://github.com/Azure/azure-sdk-for-js/blob/feature/v4/eng/pipelines/mgmt-ci.yml

This was done in the main branch in #17060

Repeating it in the feature/v4 branch as this branch is used to make updates to the older versions of the mgmt plane packages