[Identity] Fix broken link (#17072)
In this PR I’m fixing a broken link. It pointed to a feature branch, and it correctly routed to it while we were developing the feature, but after merging an deleting that branch, the link is now broken! (reasonably so). Let’s clean it up!

Thank you, @jeremymeng for letting me know about this.